Implement -c (implausbile latency filter) in ftsort
Given the option

	-c <CYCLES PER NANOSECOND>

ftsort now looks for latencies that are implausibly large and
filters them.

Whether a latency sample is "implausibly large" is determined based on
the following observation: the timestamps of earlier and later events
indicate when a processor was in preemptable mode (i.e., points in
time at which interrupts were not masked). The delta between two
consecutive points imply an upper bound on the maximum "interrupts
off" segment during the interval between the two points. No interrupt
latency should be larger than this upper bound.

Note that this option should *NOT* be required for publication-quality
benchmarks in most circumstances. In particular, this should never be
necessary when benchmarking LITMUS^RT kernels. If it becomes
necessary, then this should be considered a bug and reported.

For context, this filter was originally developed for benchmarking
SCHED_DEADLINE, in which "jobs are released" (i.e., nanosleep() calls
are completed) such that implausibly large latencies may arise if the
system becomes overloaded.